requirements.txt
alembic==1.9.2
ansible==6.2.0
ansible-core==2.13.3
bcrypt==4.0.1
cffi==1.15.1
cryptography==37.0.4
fabric==3.0.0
greenlet==2.0.1
invoke==2.0.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
packaging==21.3
paramiko==3.0.0
psycopg2-binary==2.9.5
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
python-dotenv==1.0.1
PyYAML==6.0
resolvelib==0.8.1
SQLAlchemy==1.4.46
typing_extensions==4.4.0
